+++ integrations/gitlab-v2/tests/test_client.py
@@ -2,6 +2,8 @@ from typing import Any
 from unittest import mock, IsolatedAsyncioTestCase
 from unittest.mock import patch
 
+from aiolimiter import AsyncLimiter
+
 from client import GitLabHandler
 from choices import Endpoint, Entity
 from tests.data import DATA
@@ -29,7 +31,9 @@ class ClientTest(IsolatedAsyncioTestCase):
         new_callable=mock.AsyncMock,
     )
     @patch("client.GitLabHandler.fetch_data", new_callable=mock.AsyncMock)
-    async def test_fetch_data(self, mock_fetch_data: mock.AsyncMock, mock_integration_config: mock.AsyncMock) -> None:
+    async def test_fetch_data(
+        self, mock_fetch_data: mock.AsyncMock, mock_integration_config: mock.AsyncMock
+    ) -> None:
         mock_integration_config.return_value = {
             "gitlab_url": "http://localhost:8080",
             "gitlab_token": "test_token",
@@ -59,40 +63,101 @@ class ClientTest(IsolatedAsyncioTestCase):
         self.assertEqual(fetched_issues, self.data[Entity.ISSUE.value])
         mock_fetch_data.assert_has_calls([mock.call(Endpoint.ISSUE.value)])
 
+    async def gitlab_data(self, url: str) -> list[dict[str, Any]]:
+        if Endpoint.GROUP.value in url:
+            return self.data[Entity.GROUP.value]
+        elif Endpoint.PROJECT.value in url:
+            return self.data[Entity.PROJECT.value]
+        elif Endpoint.MERGE_REQUEST.value in url:
+            return self.data[Entity.MERGE_REQUEST.value]
+        elif Endpoint.ISSUE.value in url:
+            return self.data[Entity.ISSUE.value]
+
+        raise Exception
+
     @patch(
         "port_ocean.context.ocean.PortOceanContext.integration_config",
         new_callable=mock.AsyncMock,
     )
-    @patch("client.ocean.register_raw", new_callable=mock.AsyncMock)
-    async def test_webhook_handler(self, mock_register_raw: mock.AsyncMock, mock_integration_config: mock.AsyncMock) -> None:
+    @patch("client.GitLabHandler.update_webhook", new_callable=mock.AsyncMock)
+    @patch("client.GitLabHandler.get_limiter", new_callable=mock.AsyncMock)
+    @patch("client.GitLabHandler.get_gitlab_data", new_callable=mock.AsyncMock)
+    async def test_webhook_update(
+        self,
+        mock_get_gitlab_data: mock.AsyncMock,
+        mock_get_limiter: mock.AsyncMock,
+        mock_update_webhook: mock.AsyncMock,
+        mock_integration_config: mock.AsyncMock,
+    ) -> None:
         mock_integration_config.return_value = {
             "gitlab_url": "http://localhost:8080",
             "gitlab_token": "test_token",
-            "gitlab_secret": "secretsecret",
         }
+        mock_get_limiter.return_value = AsyncLimiter(10)
+
+        mock_get_gitlab_data.side_effect = self.gitlab_data
+
+        handler = GitLabHandler()
+
+        await handler.fetch_data(Endpoint.GROUP.value)
+        mock_update_webhook.assert_called()
+
+        mock_update_webhook.reset_mock()
+        await handler.fetch_data(Endpoint.PROJECT.value)
+        mock_update_webhook.assert_not_called()
+
+        mock_update_webhook.reset_mock()
+        await handler.fetch_data(Endpoint.MERGE_REQUEST.value)
+        mock_update_webhook.assert_not_called()
+
+        mock_update_webhook.reset_mock()
+        await handler.fetch_data(Endpoint.ISSUE.value)
+        mock_update_webhook.assert_not_called()
+
+    async def gitlab_data_by_id(self, url: str) -> list[dict[str, Any]]:
+        endpoint = url.split("/")[-2]
+        if endpoint in Endpoint.GROUP.value:
+            return self.data[Entity.GROUP.value][0]
+        elif endpoint in Endpoint.PROJECT.value:
+            return self.data[Entity.PROJECT.value][0]
+        elif endpoint in Endpoint.MERGE_REQUEST.value:
+            return self.data[Entity.MERGE_REQUEST.value][0]
+        elif endpoint in Endpoint.ISSUE.value:
+            return self.data[Entity.ISSUE.value][0]
+
+        raise Exception
+
+    @patch(
+        "port_ocean.context.ocean.PortOceanContext.integration_config",
+        new_callable=mock.MagicMock,
+    )
+    @patch("client.GitLabHandler.get_gitlab_data", new_callable=mock.AsyncMock)
+    async def test_webhook_handler(
+        self,
+        mock_get_gitlab_data: mock.AsyncMock,
+        mock_integration_config: mock.AsyncMock,
+    ) -> None:
+        mock_integration_config.return_value = {
+            "gitlab_url": "http://localhost:8080",
+            "gitlab_token": "test_token",
+        }
+        mock_get_gitlab_data.side_effect = self.gitlab_data_by_id
 
         handler = GitLabHandler()
 
         await handler.group_handler(self.webhook_data[Entity.GROUP.value])
-        mock_register_raw.assert_called_with(
-            Entity.GROUP.value,
-            [{'id': 78, 'name': 'StoreCloud', 'description': 'storecloud'}],
-        )
+        mock_get_gitlab_data.assert_called()
 
+        mock_get_gitlab_data.reset_mock()
         await handler.project_handler(self.webhook_data[Entity.PROJECT.value])
-        mock_register_raw.assert_called_with(
-            Entity.PROJECT.value,
-            [{'id': 74, 'name': 'StoreCloud', 'description': 'storecloud', 'path_with_namespace': 'jsmith/storecloud'}]
-        )
+        mock_get_gitlab_data.assert_called()
 
-        await handler.merge_request_handler(self.webhook_data[Entity.MERGE_REQUEST.value])
-        mock_register_raw.assert_called_with(
-            Entity.MERGE_REQUEST.value,
-            [{'id': 99, 'title': 'MS-Viewport', 'author': {'name': 'Administrator'}, 'state': 'opened', 'created_at': '2013-12-03T17:23:34.000000Z', 'updated_at': '2013-12-03T17:23:34.000000Z', 'web_url': 'http://example.com/diaspora/merge_requests/1', 'reviewers': [{'id': 6, 'name': 'User1', 'username': 'user1', 'avatar_url': 'http://www.gravatar.com/avatar/e64c7d89f26bd1972efa854d13d7dd61?s=40&d=identicon'}], 'project_id': 1}]
+        mock_get_gitlab_data.reset_mock()
+        await handler.merge_request_handler(
+            self.webhook_data[Entity.MERGE_REQUEST.value]
         )
+        mock_get_gitlab_data.assert_called()
 
+        mock_get_gitlab_data.reset_mock()
         await handler.issue_handler(self.webhook_data[Entity.ISSUE.value])
-        mock_register_raw.assert_called_with(
-            Entity.ISSUE.value,
-            [{'id': 301, 'title': 'New API: create/update/delete file', 'web_url': 'http://example.com/diaspora/issues/23', 'description': 'Create new API for manipulations with repository', 'created_at': '2013-12-03T17:15:43.000000Z', 'updated_at': '2013-12-03T17:15:43.000000Z', 'author': {'id': 1, 'name': 'Administrator', 'username': 'root', 'avatar_url': 'http://www.gravatar.com/avatar/e64c7d89f26bd1972efa854d13d7dd61?s=40&d=identicon', 'email': 'admin@example.com'}, 'state': 'opened', 'labels': ['API'], 'project_id': 14}]
-        )
+        mock_get_gitlab_data.assert_called()