+++ integrations/gitlab-v2/CHANGELOG.md
@@ -1,9 +1,13 @@
-## 0.0.1 (2024-11-10)
+## 0.0.2 (2024-11-13)
 
 
 ### Features
 
 - Added integration for pulling of gitlab entity to port using APIs and webhook. (0.0.1)
+
+### Improvements
+
+- Move API integration to client and port integration to main. (0.0.2)
 
 
 # Changelog - Ocean - gitlab-v2
@@ -12,3 +16,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## 0.0.1 (2024-11-10)
+
+### Features
+
+- Added integration for pulling of gitlab entity to port using APIs and webhook. (0.0.1)